+++ Vibrdrome/Core/Constants/UserDefaultsKeys.swift
@@ -66,6 +66,7 @@ enum UserDefaultsKeys {
     static let showAirPlayInToolbar = "showAirPlayInToolbar"
     static let showLyricsInToolbar = "showLyricsInToolbar"
     static let showSettingsInToolbar = "showSettingsInToolbar"
+    static let nowPlayingToolbarOrder = "nowPlayingToolbarOrder"
 
     // MARK: - Discord